refactor(eo, transfers): remove transfers.store
