+++ libs/eo/transfers/src/lib/existing-transfer-agreement.ts
@@ -1,0 +1,4 @@
+export interface EoExistingTransferAgreement {
+  startDate: number;
+  endDate: number | null;
+}
